Inline Case files for NoTabsSpec
